netdata/packaging/ci: pull all git changes, prior to publishing
The way we do things within travis, we commit the packaging details withn a travis workflow.
This way when Travis goes to the next step, if we want to see that commit we need to pull.

I noticed this during v1.15.0, publishing did not see the latest stuff and produced a nightly verion due to missing tag