Use automatic dependency generation for RPM builds. (#9937)
Our current RPM spec file includes _all_ of our library dependencies as
statically listed `Requires:` lines. This, in turn, requires a very
large number of conditionals in the dependencies to account for package
naming differences between RPM distros, and also makes it a pain in the
arse to port the spec file to work with new releases of existing
distros.

This changes the spec file to just use the automatic `Requires:`
generation functionality that's already included in the RPM build
process. The net effect is that the spec file ends up much cleaner, and
it works just about anywhere with minimal modification., as well as us
not having to track anything but build dependencies manually.