Stop pulling in netcat as a mandatory dependency. (#13787)
It’s only used by a single exporter which is itself not in widespread
usage. Requiring users to install netcat themselves if they want to use
that exporter is not a significant burden, and not pulling it in
ourselves reduces our install footprint for ‘regular’ users without
really impacting usability for them.