Add -Wno-builtin-macro-redefined to compiler flags. (#17209)
This silences warnings about macros that are provided by the
preprocessor/compiler itself being redefined.

This is needed to silence warnings about _FORTIFY_SOURCE being redefined,
which are being generated as we prefer a more aggressive value for it
than what most distros provide by default. Without this change, the
build is really noisy on many systems, and also fails when run with
`-Werror`.

It is also required long-term to allow for reproducible builds without
throwing errors, as they need to redefine a number of macros provided by
the preprocessor at build time.